Make setup.py work for building with crosstool.
This patch makes setup.py sensitive to the IGNORE_SYSTEM_PATHS
environment variable. When it's set, setup.py will avoid looking in
standard places like /usr/include and /usr/lib for header files and
libraries, and will only look in the directories specified in CFLAGS
and LDFLAGS. This is important because crosstool doesn't search the
standard directories and distutils uses os.path.exists to decide
whether the compiler is going to load the file without actually
running a test build.

It also makes Python actually look in the CFLAGS and LDFLAGS paths for
headers and libraries. That's the "+ if platform == 'darwin' or
platform == 'linux2':" line, which should perhaps be more general.

Finally, because crosstool's compilers don't start with "gcc", it uses
gcc's rpath setting for any compiler that _contains_ "gcc".