Fix test_psyco to not pollute other tests.
Merely importing psyco causes the interpreter to be polluted, which will break
any tests that deal with callstack depth or tracing (and probably other things,
too). This patch sandboxes test_psyco so that the pollution is restricted to a
subprocess.

Tested with `./python -E Lib/test/regrtest.py test_psyco test_doctest test_sys`.
If that passes, the pollution has been contained.