Merge changes removing quadratic space use in long functions. ------------------------------------------------------------------------ r72411 | evancheng | 2009-05-25 23:25:46 -0700 (Mon, 25 May 2009) | 1 line
Eliminate VarInfo::UsedBlocks.
------------------------------------------------------------------------
------------------------------------------------------------------------
r72426 | jyasskin | 2009-05-26 11:27:15 -0700 (Tue, 26 May 2009) | 7 lines

LiveVariables::VarInfo contains an AliveBlocks BitVector, which has as many
entries as there are basic blocks in the function.  LiveVariables::getVarInfo
creates a VarInfo struct for every register in the function, leading to
quadratic space use.  This patch changes the BitVector to a SparseBitVector,
which doesn't help the worst-case memory use but does reduce the actual use in
very long functions with short-lived variables.

------------------------------------------------------------------------