Modify eval.cc:fast_function so more functions can use the fast path.
Since the co_flags check in fast_function was introduced in CPython r31241,
many other flags have been added to co_flags.  In particular, we added
CO_BLOCKSTACK and CO_USES_EXEC, which force many viable functions to use the
slow path.  This change checks the flags against a whitelist and a blacklist
instead.

This change results in a slight speedup, but the difference is mostly within
the noise margin.