Fix a crash in the caching imports optimization.
We need to watch the builtins dict for changes to __import__, but we can't
always rely on it to be watched already.  For example, if compilation is forced
via co_optimization, we will not have the dict.

The fix is to go ahead take a guess at the builtins dict when it's not already
being watched.